added SourceTestLoader and ConfigLoader
