+++ src/main/java/com/epam/task/test/util/ConfigLoader.java
@@ -1,0 +1,11 @@
+package com.epam.task.test.util;
+
+
+public class ConfigLoader
+{
+	public static String load() {
+
+
+		return BaseConstants.CONFIG;
+	}
+}
